src/api/requirements.txt
# requirements.txt

fastapi
python-dotenv
uvicorn
mysql-connector-python
